Refactoring: Reorganize internal modules
- Add `common::concurrent` module and move the modules shared by
  `sync`, `future` and `dash` modules there.
- Add `sync_base` module and move the modules shared by `sync` and
  `future` modules there.
- Add `common::concurrent::constants` module and move consts shared
  by `sync::base_cache` and `dash::base_cache`, etc. there.
