+++ src/common/concurrent/thread_pool.rs
@@ -6,8 +6,10 @@ use std::{collections::HashMap, sync::Arc};
 static REGISTRY: Lazy<ThreadPoolRegistry> = Lazy::new(ThreadPoolRegistry::default);
 
 #[derive(Clone, Copy, Debug, Hash, PartialEq, Eq)]
+// #[cfg_attr(any(feature = "sync", feature = "future"), derive(Debug))]
 pub(crate) enum PoolName {
     Housekeeper,
+    // #[cfg(any(feature = "sync", feature = "future"))]
     Invalidator,
 }
 
@@ -15,6 +17,7 @@ impl PoolName {
     fn thread_name_template(&self) -> &'static str {
         match self {
             PoolName::Housekeeper => "moka-housekeeper-{}",
+            // #[cfg(any(feature = "sync", feature = "future"))]
             PoolName::Invalidator => "moka-invalidator-{}",
         }
     }
