+++ src/common/time.rs
@@ -1,26 +1,13 @@
 use std::time::Duration;
 
-// target_has_atomic is more convenient but yet unstable (Rust 1.55)
-// https://github.com/rust-lang/rust/issues/32976
-// #[cfg_attr(target_has_atomic = "64", path = "common/time_atomic64.rs")]
-
-#[cfg_attr(
-    all(feature = "atomic64", feature = "quanta"),
-    path = "time/atomic_time.rs"
-)]
-#[cfg_attr(
-    not(all(feature = "atomic64", feature = "quanta")),
-    path = "time/atomic_time_compat.rs"
-)]
-pub(crate) mod atomic_time;
-
 #[cfg_attr(feature = "quanta", path = "time/clock_quanta.rs")]
 #[cfg_attr(not(feature = "quanta"), path = "time/clock_compat.rs")]
-mod clock;
+pub(crate) mod clock;
 
 pub(crate) use clock::Clock;
 
 #[cfg(test)]
+// #[cfg(all(test, feature = "sync"))]
 pub(crate) use clock::Mock;
 
 /// a wrapper type over Instant to force checked additions and prevent
@@ -41,6 +28,11 @@ impl Instant {
 
     pub(crate) fn now() -> Instant {
         Instant(clock::Instant::now())
+    }
+
+    #[cfg(feature = "quanta")]
+    pub(crate) fn inner_clock(&self) -> clock::Instant {
+        self.0
     }
 }
