+++ src/common/time/clock_quanta.rs
@@ -2,4 +2,5 @@ pub(crate) type Clock = quanta::Clock;
 pub(crate) type Instant = quanta::Instant;
 
 #[cfg(test)]
+// #[cfg(all(test, feature = "sync"))]
 pub(crate) type Mock = quanta::Mock;
